Use async communication between peer and peer_fsm.
This patch changes the communication strategy between our Peer and our
PeerFSM such that both parties use async communication between each
other. This avoids certain situations where it would otherwise be a
major hazzle to avoid the case where the peer is blocked, but the async
wants to communicate async with it.