Add rate limiting support.
This patch refactors the talla_or_peer module to be able to handle rate
limiting of the TLS connections. We move from using an active socket, to
be using a passive socket and move the socket handling intwo two
other modules: talla_or_peer_recv and talla_or_peer_send.

The new bandwidth_rate configuration entry is used to define the number
of bytes per second that is allowed to pass through the relay.